+++ Views/Models/BusinessIndexVM.cs
@@ -11,7 +11,7 @@ namespace IdleBusiness.Views.Models
         public Business Business { get; set; }
         public Entrepreneur CurrentEntrepreneur { get; set; }
         public bool HasCurrentEntrepreneurInvestedInBusiness { get; set; }
-        public float TotalInvestedAmount { get; set; }
+        public double TotalInvestedAmount { get; set; }
         public double InvestedProfits { get; set; }
         public List<Investment> Investments { get; set; }
     }
